daily

+++ tests/wasm/example/get_num_size/build.sh
@@ -1,15 +1,14 @@
 rm -rf *.wasm
+rm -rf *.bc
 
 /Applications/cheerp/bin/clang++ \
   -target cheerp-wasm \
   -cheerp-linear-heap-size=128 \
   -cheerp-linear-output=wasm \
-  -cheerp-pretty-code \
   -cheerp-no-icf \
-  -o bin.js \
-  -cheerp-global-prefix=g \
+  -o bin.js -O3 \
   -cheerp-avoid-wasm-traps \
-  main.cpp
+  main.cpp 
 
-rm -f /tmp/bin.wasm
-mv -f bin.wasm /tmp
+[ -f /tmp/bin.wasm ] && rm -f /tmp/bin.wasm || echo "skip"
+[ -f bin.wasm ] && mv -f bin.wasm /tmp || echo "skip"
